Add files via upload
This program helps you to find the paths of a given folder or folders and copy them to a delivery folder. Additionally it allows you to set the handles of the project and the extensions of the files so you don't share unwanted files. The programm will also copy the subfolders taking into account the prior specifications.